Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,7 +5,7 @@ int main() {
     scanf("%d %d %d", &a,&b,&c);
     int average;
     average=((a+b+c)/3);
-    float avg=average;
+    float avg=(float)average;
     printf("Average: %f", avg);
     return 0;
 }
